+++ CHANGELOG.md
@@ -13,6 +13,7 @@
   * Only allow one FM or Bluetooth stream to run at once
   * Change how errors are displayed on Stream create/edit modal to be more obvious
   * Fix a bug where UI content was cropped on certain mobile platforms
+  * Ensure that media device audio streams stop outputting audio when disconnected
 * System
   * Fixed a bug where support tunnel addresses would be served when querying for amplipi.local
   * Stream validation for URLs has been made more robust